chore(System): rename Proxies to HttpProxies [YTFRONT-3297]
