+++ packages/ui/src/ui/pages/system/ProxiesImpl/StatsInfo.scss
