+++ packages/ui/src/ui/pages/system/ProxiesImpl/StatsInfo.tsx
