+++ packages/ui/src/ui/pages/system/System/SystemGeneralTab.tsx
@@ -3,7 +3,7 @@ import Resources from '../Resources/Resources';
 import Masters from '../Masters/Masters';
 import SchedulersAndAgents from '../SchedulersAndAgents/SchedulersAndAgents';
 import Chunks from '../Chunks/Chunks';
-import Proxies from '../Proxies/Proxies';
+import Proxies from '../HttpProxies/HttpProxies';
 import RpcProxies from '../RpcProxies/RpcProxies';
 import Nodes from '../Nodes/Nodes';
